requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
html5lib==1.1
idna==3.4
Pillow==10.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.0.7
webencodings==0.5.1
